pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kontrol"
version = "0.1.206"
description = "Foundry integration for KEVM"
authors = [
    "Runtime Verification, Inc. <contact@runtimeverification.com>",
]

[tool.poetry.dependencies]
python = "^3.10"
kevm-pyk = { git = "https://github.com/runtimeverification/evm-semantics.git", tag = "v1.0.496", subdirectory = "kevm-pyk" }

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-quotes = "*"
flake8-type-checking = "*"
isort = "*"
mypy = "*"
pep8-naming = "*"
pytest = "^7"
pytest-cov = "*"
pytest-mock = "*"
pytest-xdist = "*"
pytest-timeout = "*"
pyupgrade = "*"

[tool.poetry.scripts]
kontrol = "kontrol.__main__:main"

[tool.poetry.plugins.kdist]
kontrol = "kontrol.kdist.plugin"

[tool.isort]
profile = "black"
line_length = 120

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true
exclude = "src/tests/integration/test-data"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.mypy]
disallow_untyped_defs = true
exclude = "src/tests/integration/test-data"
